abstract class
